+++ projects/worldskills-angular-lib-tester/src/environments/environment.ts
@@ -27,6 +27,10 @@ export const httpConfig = {
   ]
 };
 
+export const appConfig = {
+  notAuthorizedRoute: ['/not-authorized'],
+};
+
 /*
  * For easier debugging in development mode, you can import the following file
  * to ignore zone related error stack frames such as `zone.run`, `zoneDelegate.invokeTask`.