+++ projects/worldskills-angular-lib/src/lib/models/auth/user-list-view.ts
@@ -1,0 +1,11 @@
+import { UserView } from './user-view';
+export class UserListView {
+  // tslint:disable-next-line:variable-name
+  total_count: number;
+  users: UserView[];
+
+  constructor(obj?) {
+    this.total_count = obj && obj.total_count || 0;
+    this.users = obj && obj.users || [];
+  }
+}