+++ projects/worldskills-angular-lib/src/lib/models/filter-model.ts
@@ -1,0 +1,7 @@
+export class FilterModel {
+  toParams() {
+      return Object.entries(this)
+          .filter(([name, value]) => value !== null)
+          .map(([name, value]) => ({ name, value }));
+  }
+}